cluster validity measures for R

+++ src/c_common.h
@@ -72,7 +72,7 @@
 
 
 
-typedef float FLOAT_T; ///< float type we are working internally with
+typedef double FLOAT_T; ///< float type we are working internally with
 
 #ifndef INFTY
 #define INFTY (std::numeric_limits<FLOAT_T>::infinity())
@@ -82,5 +82,6 @@ typedef float FLOAT_T; ///< float type we are working internally with
 #define IS_MINUS_INFTY(x) ((x) < 0.0 && !std::isfinite(x))
 
 #define CVI_MAX_N_PRECOMPUTE_DISTANCE 10000
+
 
 #endif
